# Conflicts:
#	app/app.go
#	testutil/utils.go
#	x/delegation/keeper/delegation_state.go
#	x/deposit/keeper/deposit_test.go
#	x/operator/keeper/avs_operator_shares.go
#	x/operator/keeper/keeper.go
#	x/operator/keeper/opt.go

+++ precompiles/delegation/types.go
@@ -20,7 +20,7 @@ func (p Precompile) GetDelegationParamsFromInputs(ctx sdk.Context, args []interf
 		return nil, fmt.Errorf(cmn.ErrInvalidNumberOfArgs, 6, len(args))
 	}
 	delegationParams := &delegationtypes.DelegationOrUndelegationParams{}
-	clientChainLzID, ok := args[0].(uint16)
+	clientChainLzID, ok := args[0].(uint32)
 	if !ok {
 		return nil, fmt.Errorf(exocmn.ErrContractInputParaOrType, 0, reflect.TypeOf(args[0]), clientChainLzID)
 	}
